ml-service/requirements.txt
# FastAPI and web server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Machine Learning
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3

# Database
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Optional: MLflow for experiment tracking
# mlflow==2.9.2
